# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Metadata-Driven Enterprise Architecture Platform API
  version: 1.0.0
  description: >
    Core public REST contract for the platform. This file is intentionally a
    foundational implementation contract and SHALL be expanded alongside
    backend implementation without breaking existing documented semantics.
    Developer-facing identifiers remain English; string values may be Persian,
    and safe error messages use Persian (`fa-IR`) for the Persian-first MVP.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      operationId: login
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      operationId: refreshSession
      summary: Rotate refresh session and issue an access token
      responses:
        '200':
          description: Session refreshed; a rotated HttpOnly refresh cookie is set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Refresh session missing, invalid, expired, or reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /users/{userId}/roles:
    post:
      tags: [Identity]
      operationId: assignUserRole
      summary: Assign a global role to a user
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAssignmentRequest'
      responses:
        '200':
          description: Role assignment state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}/roles/{roleCode}:
    delete:
      tags: [Identity]
      operationId: removeUserRole
      summary: Remove a global role from a user
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: roleCode
          in: path
          required: true
          schema: {type: string, pattern: '^[A-Z][A-Z0-9_]*$'}
      responses:
        '204':
          description: Role removed or already absent
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  schemas:
    ErrorObject:
      type: object
      description: >
        Stable English machine code with a safe Persian (`fa-IR`) user-facing
        message. Clients branch on code and must not branch on translated text.
      properties:
        code:
          type: string
          example: RESOURCE_LOCKED
        message:
          type: string
          example: این مورد در مرحله قفل‌شده قرار دارد و قابل ویرایش نیست.
        details: {}
      required: [code, message]
    SuccessEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success:
          type: boolean
          const: true
        data: {}
        error:
          type: 'null'
        meta:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success:
          type: boolean
          const: false
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          type: object
          additionalProperties: true
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, minLength: 1}
        password: {type: string, minLength: 1}
    RoleAssignmentRequest:
      type: object
      required: [role_code]
      additionalProperties: false
      properties:
        role_code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
...
  responses:
    Forbidden:
      description: Effective permission is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...